ai-agent-service/requirements.txt
 # Core Dependencies
fastapi==0.128.0
uvicorn==0.40.0
pydantic==2.12.5
python-dotenv==1.2.1
python-multipart==0.0.22

# Database
pymongo==4.16.0

# LLM & Embeddings
groq==1.0.0
cohere==5.5.3

# Optional: Local Embeddings (Development Only)
# Set USE_LOCAL_EMBEDDING=true in environment to enable
# sentence-transformers==5.2.2
# torch==2.10.0
# transformers==5.0.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.2.0
beautifulsoup4==4.14.3
lxml==6.0.2

# ML & Scientific Computing
scikit-learn==1.8.0
numpy==2.4.2
scipy==1.17.0

# HTTP & Web
requests==2.32.5
httpx==0.28.1

# Required by dependencies (kept for explicit pinning)
tokenizers==0.22.2
safetensors==0.7.0
huggingface_hub==1.3.7
filelock==3.20.3
fsspec==2026.1.0
PyYAML==6.0.3
